chore: reorganized layout for better consistency across the app
