+++ app/(admin)/admin/page.tsx
@@ -1,7 +1,11 @@
+import Dashboard from '@/components/Dashboard';
+
 export default function Home() {
   return (
-    <div className='w-full h-full'>
-      <span>Admin portal coming soon!</span>
-    </div>
+    <Dashboard>
+      <div className='w-full h-full'>
+        <span>Admin portal coming soon!</span>
+      </div>
+    </Dashboard>
   );
 }
