+++ app/(site)/demo/[slug]/page.tsx
@@ -1,9 +1,5 @@
-import RecipeModule from '../../../components/RecipeModule';
+import RecipeModule from '../../../../components/RecipeModule';
 
 export default function RecipePage() {
-  return (
-    <main>
-      <RecipeModule />
-    </main>
-  );
+  return <RecipeModule />;
 }
