+++ app/(site)/layout.tsx
@@ -1,7 +1,7 @@
 import '../globals.css';
 import type { Metadata } from 'next';
-import Navigation from '../components/Navigation';
-import Footer from '../components/Footer';
+import Navigation from '../../components/Navigation';
+import Footer from '../../components/Footer';
 
 export const metadata: Metadata = {
   title: 'My Online Cookbook',
