+++ app/(site)/login/page.tsx
@@ -1,75 +1,5 @@
-import Logo from '@/app/components/Logo';
+import LoginForm from '@/components/LoginForm';
 
-export default function Example() {
-  return (
-    <>
-      <div className='flex flex-col justify-center flex-1 px-6 py-12 lg:px-8'>
-        <div className='flex flex-col items-center justify-center sm:mx-auto sm:w-full sm:max-w-sm'>
-          <Logo height='120' />
-          <h2 className='mt-10 text-2xl font-bold leading-9 tracking-tight text-center text-gray-900'>
-            Sign in to your account
-          </h2>
-        </div>
-
-        <div className='mt-10 sm:mx-auto sm:w-full sm:max-w-sm'>
-          <form className='space-y-6' action='#' method='POST'>
-            <div>
-              <label htmlFor='email' className='block text-sm font-medium leading-6 text-gray-900'>
-                Email address
-              </label>
-              <div className='mt-2'>
-                <input
-                  id='email'
-                  name='email'
-                  type='email'
-                  autoComplete='email'
-                  required
-                  className='block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-cyan-600 sm:text-sm sm:leading-6'
-                />
-              </div>
-            </div>
-
-            <div>
-              <div className='flex items-center justify-between'>
-                <label
-                  htmlFor='password'
-                  className='block text-sm font-medium leading-6 text-gray-900'
-                >
-                  Password
-                </label>
-                <div className='text-sm'>
-                  <a href='#' className='font-semibold text-cyan-600 hover:text-cyan-500'>
-                    Forgot password?
-                  </a>
-                </div>
-              </div>
-              <div className='mt-2'>
-                <input
-                  id='password'
-                  name='password'
-                  type='password'
-                  autoComplete='current-password'
-                  required
-                  className='block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-cyan-600 sm:text-sm sm:leading-6'
-                />
-              </div>
-            </div>
-
-            <button
-              type='submit'
-              className='flex w-full justify-center rounded-md bg-cyan-600 px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm hover:bg-cyan-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-cyan-600'
-            >
-              Sign in
-            </button>
-            <button
-              type='submit'
-              className='flex w-full justify-center rounded-md bg-cyan-600 px-3 py-1.5 text-sm font-semibold leading-6 text-white shadow-sm hover:bg-cyan-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-cyan-600'
-            >
-              Sign up
-            </button>
-          </form>
-        </div>
-      </div>
-    </>
-  );
+export default function Login() {
+  return <LoginForm />;
 }
